Add track events to In-App Marketplace tour
Track events were added to the following actions:
- tour started
- next step viewed
- tour dismissed
- tour completed
